feat(nfts): rename explore to nfts
