--- packages/blockchain-wallet-v4-frontend/src/layouts/Explore/NftSearch/index.ts
@@ -1,1 +1,0 @@
-export { default } from './NftSearch'
