+++ packages/blockchain-wallet-v4-frontend/src/layouts/Nfts/NftsHeader/index.ts
@@ -1,0 +1,2 @@
+export { default } from './NftsHeader'
+export { FIXED_HEADER_HEIGHT } from './NftsHeader'
