+++ packages/blockchain-wallet-v4-frontend/src/layouts/Nfts/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Nfts'
