+++ packages/blockchain-wallet-v4-frontend/src/scenes/Nfts/components/NftPageLazyLoadWrapper.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
 
-import { FIXED_HEADER_HEIGHT } from 'layouts/Explore/Header'
+import { FIXED_HEADER_HEIGHT } from 'layouts/Nfts/NftsHeader'
 
 const Wrapper = styled.div`
   height: calc(100vh - ${FIXED_HEADER_HEIGHT}px);
